firmware_file: Reorganised the firmware loading logic into a module w/ modules for each of the file types
