+++ src/flasher.rs
@@ -3,7 +3,7 @@
 // SPDX-FileContributor: Written by Mikaela Szekely <mikaela.szekely@qyriad.me>
 // SPDX-FileContributor: Modified by Rachel Mant <git@dragonmux.network>
 
-use std::io::{Read, Write};
+use std::io::Write;
 use std::path::PathBuf;
 use std::rc::Rc;
 use std::thread;
@@ -15,9 +15,9 @@ use indicatif::{ProgressBar, ProgressStyle};
 use log::{debug, error, info, warn};
 
 use crate::bmp::{self, BmpDevice, FirmwareType};
-use crate::firmware_file::FirmwareFormat;
+use crate::firmware_file::FirmwareFile;
 use crate::usb::PortId;
-use crate::{AllowDangerous, BmpParams, FlashParams, elf};
+use crate::{AllowDangerous, BmpParams, FlashParams};
 
 pub struct Firmware
 {
@@ -28,17 +28,18 @@ pub struct Firmware
 
 impl Firmware
 {
-	pub fn new<Params>(params: &Params, device: &BmpDevice, firmware_data: Vec<u8>) -> Result<Self>
+	pub fn new<Params>(params: &Params, device: &BmpDevice, firmware_file: FirmwareFile) -> Result<Self>
 	where
 		Params: BmpParams + FlashParams,
 	{
+		let firmware_data = firmware_file.firmware_data();
 		let firmware_length = firmware_data.len();
 		let firmware_length = u32::try_from(firmware_length)
 			.expect("firmware filesize exceeded 32 bits! Firmware binary must be invalid (too big)");
 
 		Ok(Self {
-			firmware_type: Self::determine_firmware_type(params, device, &firmware_data)?,
-			data: firmware_data,
+			firmware_type: Self::determine_firmware_type(params, device, firmware_data)?,
+			data: firmware_data.to_vec(),
 			length: firmware_length,
 		})
 	}
@@ -128,43 +129,6 @@ impl Firmware
 	}
 }
 
-fn intel_hex_error() -> !
-{
-	eprintln!(
-		"{} The specified firmware file appears to be an Intel HEX file, but Intel HEX files are not currently \
-		 supported. Please use a binary file (e.g. blackmagic.bin), or an ELF (e.g. blackmagic.elf) to flash.",
-		"Error:".red()
-	);
-	std::process::exit(1);
-}
-
-fn read_firmware(file_name: PathBuf) -> Result<Vec<u8>>
-{
-	let firmware_file = std::fs::File::open(file_name.as_path())
-		.wrap_err_with(|| eyre!("Failed to read firmware file {} to Flash", file_name.display()))?;
-
-	let mut firmware_file = std::io::BufReader::new(firmware_file);
-
-	let mut firmware_data = Vec::new();
-	firmware_file.read_to_end(&mut firmware_data).unwrap();
-
-	// FirmwareFormat::detect_from_firmware() needs at least 4 bytes, and
-	// FirmwareType::detect_from_firmware() needs at least 8 bytes,
-	// but also if we don't even have 8 bytes there's _no way_ this is valid firmware.
-	if firmware_data.len() < 8 {
-		return Err(eyre!("Firmware file appears invalid: less than 8 bytes long"));
-	}
-
-	// Extract the actual firmware data from the file, based on the format we're using.
-	let firmware_data = match FirmwareFormat::detect_from_firmware(&firmware_data) {
-		FirmwareFormat::Binary => firmware_data,
-		FirmwareFormat::Elf => elf::extract_binary(&firmware_data)?,
-		FirmwareFormat::IntelHex => intel_hex_error(), // FIXME: implement this.
-	};
-
-	Ok(firmware_data)
-}
-
 fn check_programming(port: PortId) -> Result<()>
 {
 	let dev = bmp::wait_for_probe_reboot(port, Duration::from_secs(5), "flash").inspect_err(|_| {
@@ -189,12 +153,12 @@ pub fn flash_probe<Params>(params: &Params, mut device: BmpDevice, file_name: Pa
 where
 	Params: BmpParams + FlashParams,
 {
-	let firmware_data = read_firmware(file_name)?;
+	let firmware_file = FirmwareFile::from_path(&file_name)?;
 
 	// Grab the the port the probe can be found on, which we need to re-find the probe after rebooting.
 	let port = device.port();
 
-	let firmware = Firmware::new(params, &device, firmware_data)?;
+	let firmware = Firmware::new(params, &device, firmware_file)?;
 
 	// If we can't get the string descriptors, try to go ahead with flashing anyway.
 	// It's unlikely that other control requests will succeed, but the OS might be messing with
